Download u-boot from SwDownload

+++ zynq_boot_bin/build_boot_bin.sh
@@ -2,12 +2,12 @@
 set -ex
 
 XSA_FILE=$1
-UBOOT_FILE=$2
+
 BUILD_DIR=build_boot_bin
 OUTPUT_DIR=output_boot_bin
 
 usage () {
-	echo "usage: $0 system_top.xsa u-boot.elf [output-archive]"
+	echo "usage: $0 system_top.xsa [output-archive]"
 	exit 1
 }
 
@@ -19,21 +19,42 @@ depends () {
 
 ### Check command line parameters
 echo $XSA_FILE | grep -q ".xsa" || usage
-echo $UBOOT_FILE | grep -q -e ".elf" -e "uboot" -e "u-boot"|| usage
 
 if [ ! -f $XSA_FILE ]; then
 	echo $XSA_FILE: File not found!
 	usage
 fi
 
-if [ ! -f $UBOOT_FILE ]; then
-	echo $UBOOT_FILE: File not found!
-	usage
-fi
-
 ### Check for required Xilinx tools (xcst is equivalent with 'xsdk -batch')
 command -v xsct >/dev/null 2>&1 || depends xsct
 command -v bootgen >/dev/null 2>&1 || depends bootgen
+
+patterns=("zed" "ccfmc_*" "ccbob_*" "usrpe31x" "zc702" "zc706" "coraz7s")
+
+carrier=$(unzip -p $XSA_FILE | grep -a "PATH_TO_FILE" | grep -oE "$(IFS='|'; echo "${patterns[*]}")")
+case  $carrier  in
+	zed)			UBOOT_FILE="u-boot_zynq_zed.elf" ;;
+	ccfmc_*)		UBOOT_FILE="u-boot_zynq_adrv9361.elf" ;;
+	ccbob_*)		UBOOT_FILE="u-boot_zynq_adrv9361.elf" ;;
+	usrpe31x)		UBOOT_FILE="u-boot-usrp-e310.elf" ;;
+	zc702)			UBOOT_FILE="u-boot_zynq_zc702.elf" ;;
+	zc706)			UBOOT_FILE="u-boot_zynq_zc706.elf" ;;
+	coraz7s)		UBOOT_FILE="u-boot_zynq_coraz7.elf" ;;
+	*)
+		echo "\n\n!!!!! Undefined carrier name for uboot selection !!!!!\n\n"
+		exit 126
+esac
+
+tool_version=$(vitis -v | grep -o "Vitis v20[1-9][0-9]\.[0-9] (64-bit)" | grep -o "20[1-9][0-9]\.[0-9]")
+if [[ "$tool_version" != "20"[1-9][0-9]"."[0-9] ]] ; then
+	echo "Could not determine Vitis version"
+	exit 1
+fi
+
+boot_partition_location=${tool_version//./_r}
+
+echo "Downloading $UBOOT_FILE ..."
+wget https://swdownloads.analog.com/cse/boot_partition_files/$boot_partition_location/$UBOOT_FILE
 
 rm -Rf $BUILD_DIR $OUTPUT_DIR
 mkdir -p $OUTPUT_DIR
